Fix package export issues compatible with ts-jest.
